Fixes to lesson layout instructions
Changed the reference to the lesson's README file -- it's not "this file" any more (this file is LAYOUT.md). 

Also removed the ">"s from the beginning of the block of text to be included in the README, to enable copying-and-pasting (and because the ~~~ should offset the quote by itself. I hope.