01-tooling: Drop remotes from fork/clone discussion
GitHub repositories know the repository they were forked from (sort of
like the usual 'origin' remote), but I'm not aware of them having
other remote-like references.

Besides dropping the "other remotes" bit, soften the 'origin' wording.
I'm not sure if GitHub uses 'origin' remotes or not to track the
parent repository, and the implementation details matter less than the
fact that there *is* a reference of some sort to the parent.