Renaming directories.
This commit does the directory-renaming part of issue #226, but
retains the sequential integer prefix for the episode file names.

Modifying or removing the numerical prefix involves changes to
the style code (so tests will still work), and possibly interacting
with the upstream Carpentries repo, which I feel should be
separated from the directory issue.